updates on the features
