+++ README.md
@@ -1,2 +1,19 @@
 # api-fb
-getting post and page data from Facebook API
+
+This app were built to help you as a data analysts to collect, transform and integrate data about social media performance (marketing) more easily.
+
+# How does it works
+
+It gets post and page data from Facebook API and save them in a Google Spreadsheet, so that you can than connect it with your BI-Tool.
+
+# Packs
+
+google-spreadsheet
+
+# Levels
+
+Page Level
+
+Post Level
+
+Ads
